refactor(KERN-1560): consolidate all modules under helpers/ folder
Move extract-filters, extract-order, and build-base-sql into helpers/
alongside aggregate-detection, naming, and ast-utils.

Structure:
  sql-to-meerkat/
  ├── helpers/
  │   ├── aggregate-detection.ts
  │   ├── ast-utils.ts
  │   ├── naming.ts
  │   ├── extract-filters.ts
  │   ├── extract-order.ts
  │   └── build-base-sql.ts
  ├── helpers.ts          (barrel re-export)
  ├── sql-to-meerkat.ts   (orchestrator)
  ├── types.ts
  └── index.ts

sql-to-meerkat.ts now imports everything from single './helpers' barrel.

work-item: KERN-1560
